backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
requests==2.31.0
python-dotenv==1.0.0

pydantic==2.5.3
pydantic-core==2.14.6

python-multipart==0.0.6
aiofiles==23.2.1
starlette==0.36.3
typing_extensions==4.9.0

google-genai==0.3.0
